version 1.2

+++ controllers/postController.js
@@ -19,10 +19,18 @@ exports.postList = async function (req, res) {
 
 //ajouter un post
 exports.postCreate = async function (req, res) {
+    const { id_User } = req;
+    // const id_Course = req.params.idCourse;
     let post = Post.build({
-        title: req.body.title, message: req.body.message, createdAt: req.body.createdAt,
-        idUser: req.body.idUser, idCourse: req.body.idCourse
+        title: req.body.title, message: req.body.message,
+        idUser: id_User, idCourse: req.body.idCourse
     })
+    if (post.title == null || post.message == null) {
+        return res.status(400).json({ 'error': 'missing parameters' })
+    }
+    if (post.title.trim() == "" || post.message.trim() == "") {
+        return res.status(400).json({ 'error': 'missing parameters' })
+    }
     await post.save()
         .then(data => {
             console.log(post.toJSON());
@@ -37,10 +45,20 @@ exports.postCreate = async function (req, res) {
 //modifier un post
 exports.postUpdate = async function (req, res) {
     if (req.params.idPost > 0) {
+        let post = Post.build({
+            title: req.body.title, message: req.body.message,
+            idUser: id_User, idCourse: req.body.idCourse
+        })
+        if (post.title == null || post.message == null) {
+            return res.status(400).json({ 'error': 'missing parameters' })
+        }
+        if (post.title.trim() == "" || post.message.trim() == "") {
+            return res.status(400).json({ 'error': 'missing parameters' })
+        }
         await Post.update(
             {
-                title: req.body.title, message: req.body.message, createdAt: req.body.createdAt,
-                idUser: req.body.idUser
+                title: post.title, message: post.message,
+                idUser: post.idUser, idCourse: post.idCourse
             },
             { where: { idPost: req.params.idPost } }
         )
@@ -57,14 +75,14 @@ exports.postUpdate = async function (req, res) {
 
 //supprimer un post
 exports.postDelete = async function (req, res) {
-    if (req.params.idPost) {
+    if (req.params.idPost > 0) {
         await Post.destroy({ where: { idPost: req.params.idPost } })
             .then(data => {
                 if (data == 0) res.status(400).json({ message: 'Post not found' });
-                else res.json(data);
+                else res.json({ message: "Post deleted" });
             })
             .catch(err => {
-                res.status(500).json({ message: "Post deleted" })
+                res.status(500).json({ message: err.message })
             })
     }
     else res.status(400).json({ message: 'Post not found' })
@@ -120,18 +138,23 @@ exports.listPostUser = async function (req, res) {
 
 //supprimer tous les posts liés à un utilisateur
 exports.deleteAllPostUser = async function (req, res) {
-    console.log(req.params.idUser);
     if (req.params.idUser) {
-        await Post.destroy({ where: { idUser: req.params.idUser } })
-            .then(data => {
-                if (data == 0) res.status(400).json({ message: 'Post not found' });
-                else res.json({ message: 'Post deleted' });
-            })
-            .catch(err => {
-                res.status(500).json({ message: err.message })
-            })
+        const userFound = await User.findOne({ where: { idUser: req.params.idUser } });
+        console.log(userFound);
+        if (userFound) {
+            await Post.destroy({ where: { idUser: req.params.idUser } })
+                .then(data => {
+                    if (data == 0) res.status(400).json({ message: 'Posts not found' });
+                    else res.json({ message: 'Post deleted' });
+                })
+                .catch(err => {
+                    res.status(500).json({ message: err.message })
+                })
+        } else {
+            return res.status(400).json({ message: 'User not found' })
+        }
     }
-    else res.status(400).json({ message: 'Post not found' })
+    else res.status(400).json({ message: 'User not found' })
 }
 
 //afficher tous les commentaires d'un post
@@ -152,3 +175,23 @@ exports.allCommentPost = async function (req, res) {
     }
     else res.status(400).json({ message: 'Post not found' })
 }
+
+//afficher tous les utilisateurs d'un post
+exports.listUsersPost = async function (req, res) {
+    if (req.params.idPost) {
+        await Comment.findAll({ attributes: ['idUser'], where: { idPost: req.params.idPost }, group: ['idUser'] })
+            .then(data => {
+                if (data != 0) {
+                    console.log("All Users:", JSON.stringify(data, null, 2));
+                    res.json(data);
+                }
+                else res.status(400).json({ message: 'Users not found' })
+
+            })
+            .catch(err => {
+                res.status(500).json({ message: err.message })
+            })
+    }
+    else res.status(400).json({ message: 'Users not found' })
+}
+
